+++ CHANGELOG.md
@@ -7,10 +7,11 @@
 - 实现 `familyhelper-finance` 模块的推送器。
   - com.dwarfeng.familyhelper.plugin.finance.handler.pusher.FamilyhelperPusher。
 
-- 实现 `notify` 模块的路由器、调度器、发送器。
+- 实现 `notify` 模块的路由器、调度器、发送器、推送器。
   - com.dwarfeng.familyhelper.plugin.notify.handler.dispatcher.GeneralDispatcherRegistry。
   - com.dwarfeng.familyhelper.plugin.notify.handler.router.PermissionRouterRegistry。
   - com.dwarfeng.familyhelper.plugin.notify.handler.sender.BuiltinSenderRegistry。
+  - com.dwarfeng.familyhelper.plugin.notify.handler.pusher.FamilyhelperPusher。
 
 - 项目建立，清除测试通过。